backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sse-starlette==2.1.3
pydantic==2.9.2
python-dotenv==1.0.1
PyYAML==6.0.2
langchain-qdrant==0.1.4
qdrant-client==1.11.3
python-multipart==0.0.12
aiofiles==24.1.0
sqlmodel==0.0.22
langchain==0.3.3
langchain-community==0.3.2
langchain-huggingface==0.1.0
langchain-text-splitters==0.3.0
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
sentence-transformers==3.1.1
rank-bm25==0.2.2
pypdf==5.0.1
docx2txt==0.8
markdown-it-py==3.0.0
huggingface-hub==0.25.2
openai==1.51.0
httpx==0.27.2
numpy==1.26.4
